backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
google-genai>=1.10.0
google-auth>=2.32.0
google-cloud-aiplatform>=1.60.0
aiohttp>=3.9.0
certifi>=2024.0.0
python-dotenv>=1.0.0
pydantic>=2.7.0
